feat(modularization): modules for freeTokens and getTokenTypeName
