+++ include/frontend/freeTokens.h
@@ -1,0 +1,8 @@
+#ifndef FREE_TOKENS_H
+#define FREE_TOKENS_H
+
+#include "frontend/lexer/core.h"
+
+void freeTokens(Token *tokens, int tokenCount);
+
+#endif // FREE_TOKENS_H
