+++ src/frontend/parser/parser.test.c
@@ -7,27 +7,7 @@
 #include "frontend/parser/core.h"
 #include "frontend/parser/printer/print_ast.h"
 #include "frontend/getTokenTypeName.h"
-
-
-/**
- * @brief Frees the memory allocated for an array of tokens.
- *
- * This function frees both the `lexeme` and `message` strings of each token, as well as the 
- * memory allocated for the array of tokens itself. It helps prevent memory leaks after token 
- * processing is complete.
- *
- * @param tokens The array of tokens to be freed.
- * @param tokenCount The number of tokens in the array.
- */
-void freeTokens(Token *tokens, int tokenCount) {
-    if (tokens != NULL) {
-        for (int i = 0; i < tokenCount; i++) {
-            free(tokens[i].lexeme);
-            free(tokens[i].message);
-        }
-        free(tokens);
-    }
-}
+#include "frontend/freeTokens.h"
 
 /**
  * @brief The entry point of the program for lexical analysis and parsing.
